requirements.txt

streamlit>=1.36
python-docx>=1.1.0
PyPDF2>=3.0.1
requests>=2.31.0
